Day 15 of 30 Days Lost in Space
Day 15 was a creative day.

We were introduced to a problem with out pin entry system. Our current system notifies the user the instant a wrong number is pressed, thus allowing an attacker to easily brute force entry into ours system.

We are left to freely develop our own system to better disguise a wrong entry, and try to fool the attacker a bit longer.

My system was to store the full 4 character pin as an array, then compare the pin to the password, and only display a wrong entry once all 4 characters were entered and tested against.